final project/requirements.txt
pillow==12.1.0
matplotlib==3.10.8
numpy==2.4.0
cycler==0.12.1
python-dateutil==2.9.0
pyparsing==3.3.1
contourpy==1.3.3
fonttools==4.61.1
kiwisolver==1.4.9
packaging==25.0
six==1.17.0